refactor: remove `dylib` from `crate-type` in Miden SDK crates
