+++ frontend-wasm/src/module/module_translation_state.rs
