+++ sdk/tx-kernel/Cargo.toml
@@ -12,7 +12,7 @@ readme.workspace = true
 edition.workspace = true
 
 [lib]
-crate-type = ["cdylib", "rlib"]
+crate-type = ["rlib"]
 
 [dependencies]
 wee_alloc = { version = "0.4.5", default-features = false}
