+++ sdk/types/Cargo.toml
@@ -12,6 +12,6 @@ readme.workspace = true
 edition.workspace = true
 
 [lib]
-crate-type = ["cdylib", "rlib"]
+crate-type = ["rlib"]
 
 [dependencies]
