+++ sdk/types/src/lib.rs
@@ -1,10 +1,5 @@
 #![no_std]
 
-#[panic_handler]
-fn my_panic(_info: &core::panic::PanicInfo) -> ! {
-    loop {}
-}
-
 // Temporary until Felt(f64) is implemented
 pub type Felt = u64;
